Scinde test-helpers en 2 fichiers
