+++ src/tests/fixtures.ts
@@ -1,9 +1,6 @@
 import { type InsertObject } from 'kysely';
-import { type User } from 'next-auth';
 
 import { type DB, kdb, sql } from '@/server/db/kysely';
-
-import { mockGetServerSession } from './setup-mocks';
 
 export async function cleanDatabase() {
   await kdb.deleteFrom('users').execute();
@@ -28,27 +25,4 @@ export async function seedTableUser(users: ReadonlyArray<Partial<InsertObject<DB
       })
     )
     .execute();
-}
-
-/**
- * Convert a number to a uuid for tests.
- *
- * Note: the number must be <= 999999999999
- */
-export function uuid(i: number): string {
-  return `${'00000000-0000-0000-0000-000000000000'.substr(0, 36 - `${i}`.length)}${i}`;
-}
-
-/**
- * Mock la session utilisateur pour les tests d'API
- */
-export function mockUserSession(user: Partial<User> | null) {
-  mockGetServerSession.mockResolvedValue(
-    user
-      ? {
-          user,
-          expires: new Date(Date.now() + 86400000).toISOString(), // 24h
-        }
-      : null
-  );
 }
